Implement a preload_descendant interface for DescendantTracker
This allow classes including DesendantTracker to define a hook
to be called when `descendants` is called in an environment that
doesn't fully eager load the application.

e.g.

```
class MyBase
  class << self
    def preload_descendants
      Dir[Rails.root.join('app/models/my_base/*.rb')].each do |path|
        require path
      end
    end
  end
end
```

While totally optional, this allows to ensure that `MyBase.descendant`
will have the same behavior in development and production.