Include migration-related rake tasks in rake -T
With their descriptions commented out these commands were not appearing
in the commands listed by rake -T, which is a shame as they are useful,
particularly during the development of more complex migrations.

Looking back at the history, it looks like they were commented out to
"Cut down even further on rake -T noise" back in 2010 (see
983815632cc1d316c7c803a47be28f1abe6698fb). I think the value of exposing
these potentially useful rake tasks to developers is greater than the
noise of three extra lines in the output of the command.