Remove obsolete or discouraged render examples
Most of these examples no longer work. The ability to render an absolute
path was removed as part of CVE-2016-0752. In Rails 6 we changed `render
file:` to only take exact paths and always to render them as raw files.
Rendering a template with its extensions _is_ supported, but I don't
think should be recommended.

This also removes `render file:` from an example where it isn't
relevant.

[ci skip]