Run :confirm callbacks around ensure_confirmation_sent
Current behavior prior to patch has `subscribe_to_channel` returning 
`true` if a confirmation is sent and `nil` if it is rejected. By putting 
an if/else block around it, we can ensure that the `:confirm` callbacks 
do not fire if the subscription will be rejected, and we get the return 
value from `ensure_confirmation_sent` if that branch is reached.

In order to get the same `nil` return from a rejection, an empty 
`return` statement is used on line 186.