Not allowed to chain ActiveSupport::Duration part methods.
Reading the issue #36644 I found a non consistent and unwanted
behavior, at least for me, when you are chaining
ActiveSupport::Duration part methods.

    1.year.year # => NoMethodError: undefined method `year' for ...
    1.month.weeks # => 18144000 days
    1.hour.minutes # => 216000 seconds

In my opinion chaining these methods should not be allowed and it
should raises the exception "NoMethodError".

This PR intercept the chain call on `method_missing` and then raises
a `NoMethodError` exception.

How it works:

    1.year.year # =>
      undefined method `year' for 1 year:ActiveSupport::Duration

    1.month.weeks # =>
      undefined method `weeks' for 1 month:ActiveSupport::Duration