Make configuration on connect optional for mysql
The configure_connection method in the abstract mysql adapter sets some
session variables that can cause proxys like proxysql to hold on to the
connection indefinetely. Furthermore there is a waitimeout set during
the configure_connection method as well, which can conflict with wait
timeouts set outside the config. Since someone can set these values in
their DB server we should have the option of being able to opt out of
this configuration entirely. This patch introduces a auto_configure key
in the mysql database config which when set to false will skip the auto
configure for that particular mysql database.