Add `exclude_nil` option to `Hash#to_query`
This option makes it possible to exclude nil values.
It's useful because when a Hash some nil values the resulting string
will be ugly, for example:

```ruby
{a: nil, b: nil, c: 'foo'}.to_query
```

With `exclude_nil` option, the resulting string is cleaner:

```ruby
{a: nil, b: nil, c: 'foo'}.to_query(exclude_nil: true)
```

It's not only cleaner but also shorter and often more expressive, because
empty value rarely makes sense.