Fix n+1 issue on ActionText attachments
The root cause of this issue is that we use globalid to query attachment
blobs, which means it can't gain any benefit from preloaded data.

So my solution is to preload those blobs and store them into a table.
And before using globalid locator to do the search, we perform a lookup
in the table with the parsed id. If it's found (which should be most of
the cases) then we just return the records. Otherwise, we perform the
search via globalid locator.