Allow passing a block for inbound mail creation
When you wanted to test a multi-part email -- for example, an email that
has both a plaintext part and an HTML part -- there wasn't a way to
easily build one without a pre-made fixture. By allowing you to pass a
block to the `create_inbound_email_from_mail` method, we unlock the
power of `Mail.new` to easily create a variety of emails, from
multi-part emails, to emails with attachments, and beyond.