Infer migrations_paths from database name
Similar to what we do for schema files, we now infer the default
`migrations_paths` for a database configuration using the database
specification name.

We still allow overriding these defaults with a `migrations_paths` key
in the database configuration file.

Additionally, update some documentation accordingly in the guides.

We've also removed `Migrator.migrations_paths` since we should be taking
the `migrations_paths` off of the config instead.

Co-authored-by: eileencodes <eileencodes@gmail.com>