Serialize symbols to strings in ImmutableString serialize method
Symbols serialize to strings when persisted:

    model.some_string = :foo
    model.save! # "foo" is persisted

This PR updates the immutable string class to serialize symbols to strings to mirror this behavior as ActiveModel::Attribute calls this serialize method to determine the return value for changed_in_place? Prior to this change this code would report that "something" had changed:

    comment = Comment.create! # (has a string "something" field)
    comment.update_column :something, :anything # persists "anything" to the "something" field of the comments table
    comment.something  # or comment.attributes
    comment.something_change # will be ["anything", "anything"], note these are `to` and `from` values, but are identical

After this PR the comment.something_change will return nil for this situation.

Fixes #36463