Change the location of a compiled rails-ujs.js
Tools like [yarn autoclean]() and [node-prune]() list `assets` as a
directory to prune, and just because `rails-ujs.js` is placed under
`lib/assets/compiled` it is deleted by them.

In my case, rails-ujs was the only package out of more than 1300
packages that got dysfunctional after running `node-prune` (or `yarn
autoclean --init && yarn install`).

A warning is added to the README document by [#31881](), but it is
easily missed.  If there is no strong reason to use the subdirectory
name `assets`, the path name should be changed so that no user
attention or action is required.

[yarn autoclean]: https://github.com/yarnpkg/yarn/blob/master/src/cli/commands/autoclean.js#L28
[node-prune]: https://github.com/tj/node-prune/blob/master/prune.go#L101
[#31881]: https://github.com/rails/rails/pull/31881