Add note about human_attribute_name symbol/string
This was always the intended API as per discussion in https://github.com/rails/rails/issues/36916, however versions before Rails 6 always passed a symbol. This means that some apps relied on symbols to be passed, this change is intended to note the change in behaviour so that apps can respond in kind.