Stop checking schema cache version before loading it
There isn't much value in this check, but it has several downsides.

The only point of it, is to save users from loading a severely outdated
cache, which arguably is a bug in the user deploy and migration process.

On the other side, if the application does implement 0-downtime deployments,
all migrations already need to be fully backward compatible, hence loading
a cache that is a couple migrations behind is much better than invalidating
it which will cause lots of schema queries to be issued.

Another issue is that queries during the boot process are inherently
problematic, because it create a hard dependency on a functional and healthy
database to allow web workers to boot, which in a outage recovery scenario
can makes things much harder.

e.g. you deploy a very slow query that overloads the database, you start
a rollback but the web workers with the sane version crash on boot because
they can't query the schema cache version.