Fix connections reaper not running after forking.
This broke in 3e2e8eeb9ea552bd4782538cf9348455f3d0e14a where it switched
to one reaper thread per process. However, the implementation will only
spawn the reaping thread if the reaping frequency has not been seen.
Since the reaping frequencies are stored in a class instance variable,
that variable is copied when the process is forked. As such, a forked
process will not spawn a new reaping thread since the class instance
variable would have contain the reaping frequency that was seen in the
parent process.

Also reverts the use of `WeakRef` introduced in
6302e56d6c1fec048f6438d9f1ac6a8cfaed7eb9 since we properly deregisters
connection pools from the reaper upon discard/disconnect.