Allow reaper thread exit if all connections closed
Previously, once we spawned a reaper thread we would keep it around
forever. This was mostly fine because basically nothing (other than
tests) is likely to configure this time and so it should only have been
one thread.

That said, we might as well clean it up properly if all connection pools
are removed.