Support extra stuff in the Content-Type header via Mime registration
I'm not sure a better way to describe this, but we need to set extra
stuff in the Content-Type header and we're doing it via Mime
registration so that we can have the "html_fragment" method on the
respond_to type.  This worked in Rails 5.2, but it's broken in 6.0

In the future, I think we should probably give `register` a block that
takes the response object so we can do more interesting and complex
stuff without doing weird string manipulation.  But this patch fixes the
issue we have today.