Output routes definition in json format by `rails routes -o json`
The code outputs routing definition by JSON format to stdout in
"rails routes --format=json" command.

Result of "rails routes" command is easy to read for human but it's hard
to parse and we have no way to use public API to get an easily parsable
routing format of the application.