Workaround for kwargs 2.7 and 2.6 incompatibility
Current Ruby 2.7 implementation behaves differently from 2.6 (by design)
where receiving an empty keyword arguments.

Thus, for now we need this if branch everywhere we're asserting such a method call
in order for the CI to pass.

We could revert this workaround if Ruby 2.7 decides to revert back to the 2.6 behavior,
or maybe we'll need to fix Minitest::Mock once the 2.7 spec became stable.