concerning learns how to prepend the concern
In rare instances, a module needs to be prepended into the ancestor
hierarchy, not simply included. In cases where this is necessary, and it
is desirable to use an inline concern, it would be beneficial to be able
to instruct `concerning` that the concern should be prepended.

This is now possible, by providing `prepend: true` kwarg to
`concerning`. (It is false by default.)