Add create database actionable error
This commit includes a create database action for
`ActiveRecord::NoDatabaseError`.
I noticed that `rails db:prepare` is included in `bin/setup` but I found
this useful when I run `rails new`. The command will install all the
dependencies, but it doesn't create the database, so I often forget to
do it.
I guess we could also include the database creation inside `rails new`,
but I wanted to open this to know what you think about it.

Note: I just opened this with a simple implementation to get some
feedback. If this turns out to be wanted, I'll include tests and discuss
the multi-database case.