Avoid setting up file watcher on cache_classes
Previously, if cache_classes was true we would build a file_watcher and
add it to reloaders though it would never be used. This wasted some
memory for FileUpdateChecker to maintain an array of all watched files,
and added slightly to boot time.

Instead we should avoid building a file watcher it when cache_classes is
true.