Deprecate `current_config` and `current_config=`
Uses of `current_config` and `current_config=` were removed when
implementing multiple datatabase Rails tasks. This means they are no
longer used internally. However, since it is a documented method it may
be used in gems or applications. This change deprecates these methods so
that we can remove them in 6.2.

Co-authored-by: John Crepezzi <john.crepezzi@gmail.com>