Fix ignore fixture tests causing random AR failures
Since #36303 we allow to ignore specific fixtures from being loaded, but
because that introduced a new fixture `other_books` that uses the same
underlying model `Book`, sometimes depending on the run order the tests
might end up not loading the actual `books` fixtures thinking they are
already cached/loaded. Disabling the transactional fixtures for that
test that was introduced above for the ignore fixtures feature fixes
that.

Closes #37244.
Closes #37245.
Thanks @yahonda for finding a way to reproduce these failures.