Prefer one liners OS requirement install commands
I was getting some problems when running `bundle install` on a fresh
clone. It was due to some extension building errors that usually point
out to missing OS dependencies.

So I run "rg apt-get" hoping to find some suggestion to install
requirements for my OS. However, the result was misleading, because
dependencies are split across several lines, so I missed many of them.

I think a one liner is better and it's also slightly easier to
copy-paste.