Resolve Hash with `url` key as a `UrlConfig`
In `DatabaseConfigurations#configs_for`, `Hash` configs that contain a
`:url` key resolve to a `UrlConfig`.  Since we weren't mirroring that behavior
in `Resolver`, we had to add explicit handling for `url` keys to the
`HashConfig` implementation.

Now we're going to always resolve these hashes as `UrlConfig` and can
remove the extra implementation of URL handling from `HashConfig`.

Of course, this does introduce some duplication between
`DatabaseConfigurations` and `Resolver` which we plan to address in a
future PR.

Co-authored-by: eileencodes <eileencodes@gmail.com>