Revert "Make all reads on configuration_hash use methods"
Instead of introducing individual methods for every property on `configuration_hash`,
we're aiming to maintain that we only define methods for properties
related to how Rails will _use_ a connection (adapter, spec, env,
replica?) and not the database connection information itself.

To that end, we're reverting #37280 which introduced methods specific to
connection properties like `pool`, `idle_timeout`, and `database`.

This reverts commit b8b2d406591349d24d2ced6d9ef67395166f0c88.

Co-authored-by: eileencodes <eileencodes@gmail.com>